requirements.txt
kubernetes
jinja2
Flask
PyYAML
jinja2==3.1.3
